Implement HomeClient component and integrate Amplitude analytics
- Introduced a new `HomeClient` component to manage the homepage layout and display popular mixes.
- Replaced the previous homepage structure with a server-side rendered approach to fetch popular mixes from Supabase.
- Added `AmplitudeAnalytics` component for improved analytics tracking.
- Updated `RecentMixes` component to accept initial mixes as props, optimizing data fetching.
- Implemented a new API route to increment mix view counts atomically.

These changes enhance the homepage functionality and analytics capabilities, providing a better user experience.
