+++ src/app/layout.tsx
@@ -6,6 +6,8 @@ import Navigation from "./components/Navigation";
 import { AudioProvider } from "./contexts/AudioContext";
 import { UniversalPlayer } from "./components/UniversalPlayer";
 
+import AmplitudeAnalytics from "./components/AmplitudeAnalytics";
+
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
@@ -21,33 +23,7 @@ export default function RootLayout({
   return (
     <html lang="en">
       <head>
-      <Script
-          strategy="afterInteractive"
-          src="https://cdn.amplitude.com/script/f0e77ea0c5ad241d6508e1c7a0c354e9.js"
-        />
-        <Script
-          id="amplitude-init"
-          strategy="afterInteractive"
-          dangerouslySetInnerHTML={{
-            __html: `
-              window.amplitude.add(window.sessionReplay.plugin({sampleRate: 1}));
-              window.amplitude.init('f0e77ea0c5ad241d6508e1c7a0c354e9', {
-                "fetchRemoteConfig": true,
-                "autocapture": {
-                  "attribution": true,
-                  "fileDownloads": true,
-                  "formInteractions": true,
-                  "pageViews": true,
-                  "sessions": true,
-                  "elementInteractions": true,
-                  "networkTracking": true,
-                  "webVitals": true,
-                  "frustrationInteractions": true
-                }
-              });
-            `,
-          }}
-        />
+        <AmplitudeAnalytics />
         <Script
           strategy="afterInteractive"
           src="https://www.googletagmanager.com/gtag/js?id=G-P8692YZSLG"
