+++ src/app/mix/[id]/page.tsx
@@ -4,7 +4,7 @@ import { MixPlayer } from './MixPlayer';
 import { notFound } from 'next/navigation';
 import { Metadata } from 'next';
 
-export const revalidate = 0;
+export const revalidate = 60;
 
 type Params = Promise<{ id: string }>;
 
@@ -79,12 +79,6 @@ async function getMixData(id: string) {
     .from('likes')
     .select('*', { count: 'exact', head: true })
     .eq('mix_id', id);
-
-  // Update play count
-  await supabase
-    .from('mixes')
-    .update({ play_count: (mix.play_count || 0) + 1 })
-    .eq('id', id);
 
   return {
     mix,
