+++ src/instrumentation-client.ts
@@ -10,3 +10,4 @@ if (posthogKey) {
   })
 }
 
+
